requirements.txt
# requirements.txt
streamlit==1.31.0
groq==0.9.0
httpx==0.24.1
chromadb==0.4.22
sentence-transformers==2.3.1
todoist-api-python==2.1.4
python-dotenv==1.0.0
python-telegram-bot==20.7